DETECTIFY/requirements.txt
streamlit
opencv-python
face-recognition
PyYAML
numpy
pandas
requests
Pillow